src_Tray/requirements.txt
pystray
Pillow
psutil